pyproject.toml
[tool.poetry]
name = "simple-crud"
version = "0.1.0"
description = ""
authors = ["Kartheek <karthikasasanka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
tortoise-orm = "^0.16.15"
asyncpg = "^0.21.0"
pydantic = "^1.6.1"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
pytest = "^6.0.2"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
